backend/requirements.txt
fastapi==0.110.0
starlette==0.36.3
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.3.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pgvector==0.3.0
redis[hiredis]==5.0.4
celery==5.4.0
python-dotenv==1.0.1
httpx==0.27.0
google-generativeai>=0.8.3
PyGithub==2.3.0
python-gitlab==4.6.0
alembic==1.13.1
python-multipart==0.0.9
aiofiles==23.2.1
tenacity==8.3.0
structlog==24.2.0
